Create FizzBuzz-Naive.py
Example uses old fashioned way of solving the FizzBuzz problem by checking both divisors of 3 and 5 in order to add "FizzBuzz", than checking 3 by itself to add "Fizz", followed by checking 5 in order to add "Buzz", ending with simply printing the number as a string. This approach, while working, does not account for divisors of 7 or others.